# Iterators: re-implement entrySortIterator using loserTree for performance