--- a/pkg/util/loser/tree.go
+++ b/pkg/util/loser/tree.go
@@ -2,13 +2,11 @@
 
 package loser
 
-type Value interface{}
-
 type Sequence interface {
 	Next() bool // Advances and returns true if there is a value at this new position.
 }
 
-func NewMerge[E Value, S Sequence](nElements int, maxVal E, at func(S) E, less func(E, E) bool, close func(S)) *Tree[E, S] {
+func NewMerge[E any, S Sequence](nElements int, maxVal E, at func(S) E, less func(E, E) bool, close func(S)) *Tree[E, S] {
 	t := Tree[E, S]{
 		maxVal: maxVal,
 		at:     at,
@@ -40,7 +38,7 @@ func (t *Tree[E, S]) Close() {
 // A loser tree is a binary tree laid out such that nodes N and N+1 have parent N/2.
 // We store M leaf nodes in positions M...2M-1, and M-1 internal nodes in positions 1..M-1.
 // Node 0 is a special node, containing the winner of the contest.
-type Tree[E Value, S Sequence] struct {
+type Tree[E any, S Sequence] struct {
 	maxVal E
 	at     func(S) E
 	less   func(E, E) bool
@@ -48,7 +46,7 @@ type Tree[E Value, S Sequence] struct {
 	nodes  []node[E, S]
 }
 
-type node[E Value, S Sequence] struct {
+type node[E any, S Sequence] struct {
 	index int // This is the loser for all nodes except the 0th, where it is the winner.
 	value E   // Value copied from the loser node, or winner for node 0.
 	items S   // Only populated for leaf nodes.