**What this PR does / why we need it**:

[Draft PR for comment at this time]

This implementation uses a data structure called Loser Tree, also known as Tournament Tree, based on Knuth, "Sorting and Searching" section 5.4.1.

I couldn't find an existing Loser Tree in Go, so I wrote my own; the file is 132 lines long.  It uses generics in the hope it can also be applied elsewhere.

One benchmark result: 
```
name                 old time/op    new time/op    delta
SortIterator/sort-4    3.78ms ± 4%    2.66ms ± 3%  -29.54%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
SortIterator/sort-4     319kB ± 0%      14kB ± 0%  -95.52%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
SortIterator/sort-4       104 ± 0%         5 ± 0%  -95.19%  (p=0.008 n=5+5)
```

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- NA Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- NA Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`