@cyriltovena 2023-01-31
Do you think this can be applicable to the `mergeEntryIterator` ? there's also alternative iterators for `sample` iteration.