@bboreham 2023-01-31
> Do you think this can be applicable to the mergeEntryIterator ?

Yes; I started with what looked like the simplest one.