@bboreham 2023-02-22
I believe I have incorporated all feedback, and have marked this ready for review.
`mergeEntryIterator` is nearly done, but we could go ahead with this if you like.